requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
ccxt==4.5.4
pandas==2.3.1
jinja2==3.1.6
python-dotenv==1.1.1
websockets==13.1
certifi==2026.2.2
